Refactor: move first_conflicting_index() to RaftState
